refactor du nom de l'application et des packages
