+++ app/src/main/java/polytech/android/LGLPDR/AndroidApp/HomeActivity.java
@@ -1,4 +1,4 @@
-package fr.greyl.robin.AndroidTP;
+package polytech.android.LGLPDR.AndroidApp;
 
 import android.content.Intent;
 import android.os.Bundle;
