+++ app/src/main/res/layout/activity_home.xml
@@ -4,7 +4,7 @@
     android:layout_height="match_parent"
     xmlns:tools="http://schemas.android.com/tools"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context="fr.greyl.robin.AndroidTP.HomeActivity">
+    tools:context="polytech.android.LGLPDR.AndroidApp.HomeActivity">
 
     <Button
         android:id="@+id/beer_app"
