+++ app/src/main/res/layout/activity_selection.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:padding="10dp"
-    tools:context="fr.greyl.robin.AndroidTP.SelectionActivity">
+    tools:context="polytech.android.LGLPDR.AndroidApp.SelectionActivity">
 
     <TableLayout
         android:id="@+id/layoutOffline"
